backend/requirements.txt
torch
transformers
sentence-transformers
accelerate
bitsandbytes
scikit-learn
pandas
selenium
webdriver_manager
bs4
fitz
spacy
PyMuPDF
openai
huggingface_hub[hf_xet]
argparse
psycopg2-binary
pymupdf>=1.23
peft
flask